+++ packages/jinajs/src/jinaClient.ts
@@ -61,8 +61,29 @@ export class JinaClient<IRequestBody = AnyObject ,IResponseData = AnyObject> {
   ): Promise<{ results: SimpleResults[]; queries: SimpleQueries }> {
     const requestBody = await this.serializeRequest(documents);
     console.log("request body:", requestBody);
-    const response = await this.client.post("search", requestBody);
+    const response = await this.client.post("search", requestBody, {
+      headers: {
+        'content-type': ''
+      }
+    });
     console.log("response:", response);
     return this.serializeResponse(response);
   }
+
+  async searchWithParameters(
+    documents: RawDocumentData[],
+    parameters: AnyObject
+  ): Promise<{ results: SimpleResults[]; queries: SimpleQueries }> {
+    console.log(parameters)
+    const requestBody = await this.serializeRequest(documents);
+    console.log("request body:", requestBody);
+    const response = await this.client.post("search", requestBody, {
+      headers: {
+        'content-type': ''
+      }
+    });
+    console.log("response:", response);
+    return this.serializeResponse(response.data);
+  }
+  
 }